feat(hokm): ادغام بازطراحی بصری کاربر (هویت ایرانی: شمسه/اسلیمی/گره‌چینی) + رفع ناسازگاری‌ها
- فایل‌های آپلودشده در مسیر اشتباه (hokm/ci) به جای درست خودشان منتقل شدند:
  • core/app_theme.dart — پالت لاجورد/فیروزه/طلا با حفظ توکن‌های قدیمی
  • game/decor/persian_motifs.dart (جدید) — کتابخانهٔ نقش‌مایه‌های ایرانی
  • game/components/card_component.dart — سایهٔ سه‌لایه، هالهٔ تنفسی، دیم، برق برنده
  • game/animations/motion.dart — منحنی‌های HokmCurves + flyTo/flip/shake
  • game/table/table_theme.dart — رنگ‌های medallion/glaze/lightWash با سازگاری کامل
  • game/effects/{turn_indicator,win_glow}.dart — شمسهٔ چرخان، کمان نوری، جرقه
- چهار فایل تکراریِ « (1).dart» حذف شدند.
- رفع خطاهای کامپایل: خروجی num از clamp() به double تبدیل شد
  (card_component / win_glow / persian_motifs).
- CardRenderer: پارامترهای جدید foilPhase (برق طلایی روبروی کارت برنده)
  و shimmer (برق ملایم پشت کارت) اضافه شد.
- hokm_game: فلگ winner هنگام جمع‌کردن دست ست/پاک می‌شود.
- CI: بازسازی بهبودهای ورک‌فلو — concurrency (لغو ران قدیمی)،
  timeout-minutes، annotationهای ::error:: برای خطاهای analyze/test،
  و خلاصهٔ فشردهٔ خطا/هشدار در کامنت تشخیصی.

نکته: hokm/ci/hokm-flutter.yml باید به .github/workflows/hokm-flutter.yml
همگام شود (بات اجازهٔ نوشتن workflows را ندارد).

+++ hokm/lib/game/decor/persian_motifs.dart
@@ -209,15 +209,15 @@ abstract final class PersianMotifs {
 
   /// گرادیانِ طلاییِ متحرک؛ [phase] بین ۰ و ۱ برقِ عبوری را جابه‌جا می‌کند.
   static Shader goldShader(Rect rect, {double phase = 0.35}) {
-    final p = phase.clamp(0.0, 1.0);
+    final p = phase.clamp(0.0, 1.0).toDouble();
     return LinearGradient(
       begin: Alignment.topLeft,
       end: Alignment.bottomRight,
       stops: [
         0.0,
-        (p - 0.16).clamp(0.02, 0.9),
-        p.clamp(0.05, 0.95),
-        (p + 0.16).clamp(0.1, 0.98),
+        (p - 0.16).clamp(0.02, 0.9).toDouble(),
+        p.clamp(0.05, 0.95).toDouble(),
+        (p + 0.16).clamp(0.1, 0.98).toDouble(),
         1.0,
       ],
       colors: const [
